Refactor lib/js folder
