+++ src/assets/js/lib/error.js
@@ -1,0 +1,10 @@
+export const converErrorArrToObj = (errArr) => {
+  const errorObj = {};
+
+  errArr.forEach((err) => {
+    if (!Array.isArray(errorObj[err.param])) errorObj[err.param] = [];
+    errorObj[err.param].push(err.msg);
+  });
+
+  return Object.fromEntries(Object.entries(errorObj).map(([key, val]) => [key, val.join('; ')]));
+};
