+++ src/assets/js/lib/file.js
@@ -1,0 +1,1 @@
+export const getImgFromFile = (file) => URL.createObjectURL(file);
