+++ src/assets/js/lib/string.js
@@ -1,0 +1,1 @@
+export const capitalize = (str = '') => (str.length > 0 ? `${str[0].toUpperCase()}${str.slice(1)}` : '');
